+++ mlflow/utils/autologging_utils.py
@@ -5,6 +5,7 @@ import warnings
 
 import mlflow
 
+
 INPUT_EXAMPLE_SAMPLE_ROWS = 5
 ENSURE_AUTOLOGGING_ENABLED_TEXT = (
     "please ensure that autologging is enabled before constructing the dataset."
